+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Refactor `MEDIA_CATEGORY` enum
 
 ## 0.17.0
 * Configure various algorithm weight presets that can be selected via `PresetWeights` object